вынести поиск и рендер карточек в отдельный файл
